# internal/transport/http/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "User Service API",
    "description": "API для управления пользователями",
    "version": "1.0.0",
    "contact": {
      "name": "WebCraftersGH",
      "url": "https://github.com/WebCraftersGH/User-service"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT токен для авторизации"
      }
    },
    "schemas": {
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "username": {
            "type": "string",
            "example": "john_doe"
          },
          "fio": {
            "type": "string",
            "example": "John Doe"
          },
          "bio": {
            "type": "string",
            "example": "Software developer"
          },
          "sex": {
            "type": "string",
            "enum": ["male", "female", "other"],
            "example": "male"
          },
          "birthday": {
            "type": "string",
            "format": "date",
            "example": "1990-01-01"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "user not found"
          }
        }
      },
...
    },
    "responses": {
...
      "NotFoundError": {
        "description": "Ресурс не найден",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "user not found"
            }
          }
        }
      },
      "BadRequestError": {
        "description": "Некорректный запрос",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "parseError": {
                "value": {"error": "parse uuid error"}
              },
              "jsonError": {
                "value": {"error": "invalid json input json data"}
              }
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Внутренняя ошибка сервера",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "internal error"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/users/{uuid}": {
      "get": {
        "summary": "Получить пользователя по ID",
        "description": "Возвращает информацию о пользователе по его UUID",
        "tags": ["Users"],
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "description": "UUID пользователя",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "responses": {
          "200": {
            "description": "Пользователь найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
...
  "security": []
}